Remove usage of data ranges for water sampling
