+++ packages/website/src/common/SiteDetails/index.tsx
@@ -13,7 +13,7 @@ import times from 'lodash/times';
 
 import { useDispatch, useSelector } from 'react-redux';
 import { oceanSenseConfig } from 'constants/oceanSenseConfig';
-import type { Site, LatestDataASSofarValue, Sources } from 'store/Sites/types';
+import type { Site, LatestDataASSofarValue } from 'store/Sites/types';
 import { SurveyListItem, SurveyPoint } from 'store/Survey/types';
 import {
   forecastDataRequest,
@@ -46,7 +46,6 @@ import WaterSamplingCard from './WaterSampling';
 import { styles as incomingStyles } from './styles';
 import LoadingSkeleton from '../LoadingSkeleton';
 import playIcon from '../../assets/play-icon.svg';
-import HUICard from './HUICard';
 
 const sondeMetrics: (keyof LatestDataASSofarValue)[] = [
   'odoConcentration',
@@ -76,10 +75,6 @@ const SiteDetails = ({
   const [hasSondeData, setHasSondeData] = useState<boolean>(false);
   const [hasSpotterData, setHasSpotterData] = useState<boolean>(false);
   const [hasHUIData, setHasHUIData] = useState<boolean>(false);
-  const [
-    availableSourcesForSelectedDates,
-    setAvailableSourcesForSelectedDates,
-  ] = useState<Sources[]>([]);
   const latestData = useSelector(latestDataSelector);
   const forecastData = useSelector(forecastDataSelector);
   const isMobile = useMediaQuery(theme.breakpoints.down('xs'));
@@ -113,18 +108,15 @@ const SiteDetails = ({
       const hasSpotter = Boolean(parsedData.bottomTemperature);
       const hasSonde =
         sondeMetrics.filter((x) => Boolean(parsedData[x])).length >=
-          MINIMUM_SONDE_METRICS_TO_SHOW_CARD ||
-        availableSourcesForSelectedDates.includes('sonde');
-      const hasHUI =
-        latestData.some((x) => x.source === 'hui') ||
-        availableSourcesForSelectedDates.includes('hui');
+        MINIMUM_SONDE_METRICS_TO_SHOW_CARD;
+      const hasHUI = latestData.some((x) => x.source === 'hui');
 
       setHasSondeData(hasSonde);
       setHasSpotterData(hasSpotter);
       setHasHUIData(hasHUI);
       setLatestDataAsSofarValues(parsedData);
     }
-  }, [availableSourcesForSelectedDates, forecastData, latestData]);
+  }, [forecastData, latestData]);
 
   const { videoStream } = site || {};
 
@@ -143,15 +135,7 @@ const SiteDetails = ({
           (() => {
             if (hasHUIData) {
               return (
-                <HUICard
-                  data={{
-                    salinity: latestDataAsSofarValues.salinity,
-                    turbidity: latestDataAsSofarValues.turbidity,
-                    nitratePlusNitrite:
-                      latestDataAsSofarValues.nitratePlusNitrite,
-                    ph: latestDataAsSofarValues.ph,
-                  }}
-                />
+                <WaterSamplingCard siteId={site.id.toString()} source="hui" />
               );
             }
             if (hasSondeData) {
@@ -318,7 +302,6 @@ const SiteDetails = ({
           site={site}
           selectedSurveyPointId={selectedSurveyPointId}
           surveys={surveys}
-          setParentAvailableSources={setAvailableSourcesForSelectedDates}
         />
         <Surveys site={site} />
       </Box>
